Set up Visual Studio Code with Unity, and completed basic player movement implementation
The main feature is in PlayerMover.cs, which is my first attempt at creating a grid based movement mechanic for the player of the game.
I also created a test scene for this called MovementTest.
While this approach seems functionally correct, it doesn't feel correct.
Next area to improve is to try making it so that a single frame input sets your destination for the next tile, but this can be canceled by a reverse single frame input.