Pushed a basic box implementation which allows the player to pick up and throw boxes
Pressing E will allow you to pick up or place a box in the adjacent tile the player is facing.
Holding E while having a box will charge up the throw, so that when you release, the box will be thrown a corresponding number of tiles in the direction you're facing.